split tests by sample so they are in their own files
